engine/pyproject.toml
[project]
name = "engine"
version = "0.1.0"
description = "Engine-level orchestration for Risk League: league logic, daemons, shared infrastructure"
authors = [{ name = "Zach McMullen", email = "zachmcmullen04@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "nba_api",
  "pandas",
  "sqlalchemy",
  "psycopg2",
  "python-dotenv",
  "scikit-learn",
  "numpy",
  "requests",
  "apscheduler",
  "black",
  "sqlacodegen",
  "MLB-StatsAPI",
  "asyncio",
  "mypy",
  "redis",
  "python-socketio[asyncio_client]"
]

[tool.setuptools]
packages = [
  "nba",
  "nba.tests",
  "system",
  "shared",
  "shared.prop_generation",
  "shared.tests",
  "mlb",
  "mlb.tests"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
